refactor: task list and form
